requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

Required packages for BCI

PyQt5_sip == 12.8.1
PyWavelets == 1.1.1
joblib == 0.16.0
mysql_connector == 2.2.9
numpy == 1.19.1
pandas == 1.1.1
pygame == 2.0.1
scikit_learn == 0.23.1
scipy == 1.5.2
tabulate == 0.8.9
websocket_client == 0.57.0
